requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environ==1.0
gunicorn==22.0.0
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
sqlparse==0.5.2
tzdata==2024.2
